+++ src/App.tsx
@@ -11,6 +11,7 @@ import {
   PttListener,
   QuitConfirmListener,
 } from '@/features/system'
+import { UpdaterBoot } from '@/features/updater'
 import { Home } from '@/routes/Home'
 import { StyleGuide } from '@/routes/StyleGuide'
 import { readWindowStyleBootCache } from '@/stores/settingsStore'
@@ -50,6 +51,7 @@ function App() {
       <PttListener />
       <QuitConfirmListener />
       <PomodoroNotifyListener />
+      <UpdaterBoot />
       <BrowserRouter>
         <ChromeAwareShell>
           <ErrorBoundary surface="routes">